固定小数点型を作る (#23)
* Protos

* Work fixed point

* Work fixed point

* refactoring

* refactoring

+++ src/cnn/cnn_main.cpp
@@ -2,11 +2,12 @@
 
 #include "util/read_data.hpp"
 #include "cnn/cnn.hpp"
+// #include "util/types.hpp"
 #include "half.hpp"
 
 using half_float::half;
 int main() {
-  CNN<float>::run();
+  CNN<half>::run();
   return 0;
 }
